requirements.txt
aiosqlite
alembic
bcrypt
asyncpg
fastapi_pagination
fastapi_jwt_auth
httpx
psycopg2-binary
pydantic[email]
python_dotenv
pytest
pytest_asyncio
pytest_mock
sqlalchemy
sqlalchemy_utils
stackprinter
uvicorn

